\name{GroAgeData}
\alias{GroAgeData}
\title{
Class Attribute of LifeHist Data Object
}
\description{
To be used by plot.GroAgeData() to examine raw length, age data before modeling.
}
\usage{
GroAgeData(x, ...)
}
\arguments{
  \item{x}{
A list object coming from as.GroAgeData().
}
  \item{\dots}{
Not used.
}
}
\value{
A class attribute
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See example for as.GroAgeData().
}
\keyword{ ~classes }

