\name{GroAgeExp}
\alias{GroAgeExp}
\title{
Class Attribute of LifeHist Exploratory Model Object
}
\description{
To be used by plot.GroAgeExp() to examine the quality of initial parameter values
before fitting a growth model.
}
\usage{
GroAgeExp(x, ...)
}
\arguments{
  \item{x}{
A list object coming from groageexp().
}
  \item{\dots}{
Not used.
}
}
\value{
A class attribute
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples for plot.AgeGroExp()
}
\keyword{ ~classes }
