\name{GroAgeMod}
\alias{GroAgeMod}
\title{
Class Attribute of LifeHist Fitted Model Object
}
\description{
To be used by plot.GroAgeMod() to examine results of fitting a growth model.
}
\usage{
GroAgeMod(x, ...)
}
\arguments{
  \item{x}{
A list object coming from GroAgeFit().
}
  \item{\dots}{
Not used.
}
}
\value{
A class attribute
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples for plot.AgeGroMod()
}
\keyword{ ~classes }