\name{ksbream}
\alias{ksbream}
\docType{data}
\title{
King Soldier Bream Individual Biological Data
}
\description{
King soldier bream individual-based data from the Arabian Gulf sampled from the
catch of two artisanal fishing fleets.
}
\usage{data("ksbream")}
\format{
  A data frame with 797 observations on the following 14 variables.
  \describe{
    \item{\code{CollectionDate}}{Date of capture}
    \item{\code{TL.mm}}{Total length in mm}
    \item{\code{FL.mm}}{Fork length in mm}
    \item{\code{SL.mm}}{Standard length in mm}
    \item{\code{TW.g}}{Body weight in g}
    \item{\code{Ewt.g}}{Eviscerated body weight in g}
    \item{\code{Liver.Wt.g}}{Liver weight in g}
    \item{\code{GW.g}}{Gonad weight in g}
    \item{\code{Sex}}{1 for males, 2 for females}
    \item{\code{Mature}}{1 for mature, 0 for juvenile}
    \item{\code{Age.yr}}{Age in years}
    \item{\code{MaturityStage}}{Macroscopic maturity stages}
  }
}
\examples{
data(ksbream)
}
\keyword{datasets}
