\name{MatData}
\alias{MatData}
\title{
Class Attribute of LifeHist Data Object
}
\description{
To be used by plot.MatData() to examine data versus model predictions.
}
\usage{
MatData(x)
}
\arguments{
  \item{x}{
An object of class MatData, generated by function MatCount().
}
}
\details{
This generic function will be used to give a class to the four components
of the list output by function MatCount().
}
\value{
A class attribute.
}
\author{
Ruben H. Roa-Ureta
}
\note{
Objects of class MatData are plotted for exploration of parameter values (initial
or maximum likelihood) in 2- or 3-parameter logistic regression
}
\examples{
#See example in man pages for plot.MatData().
}