\name{plot.GroAgeData}
\alias{plot.GroAgeData}
\title{
Exploratory Analysis of Agen and Length Individual Data
}
\description{
Allows examining the relation between age and length in order to estimate
individual growth models, and optionally other individual processes such
as the relation between length and body weight and the timing of organ growth
along the annual season.
}
\usage{
\method{plot}{GroAgeData}(x, ...)
}
\arguments{
  \item{x}{
An object of class 'GroAgeData' produced by function as.GroAgeData().
}
  \item{\dots}{
Not used.
}
}
\value{
A multipanel plot.
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples in help pages for function as.GroAgeData()
}
\keyword{ ~iplot }