\name{plot.GroAgeExp}
\alias{plot.GroAgeExp}
\title{
Exploratory Modeling of Individual Growth with Age and Length Data
}
\description{
Evaluate and refine the goodness of initial parameter values before fitting
growth models to data.
}
\usage{
\method{plot}{GroAgeExp}(x, ...)
}
\arguments{
  \item{x}{
An object of class 'GroAgeExp' created by function groageexp().
}
  \item{\dots}{
Not used.
}
}
\value{
A multipanel plot.
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples for function groageexp()
}
\keyword{ ~iplot }