\name{plot.GroAgeMod}
\alias{plot.GroAgeMod}
\title{
Examination of Results from Fitting Growth Models
}
\description{
After model fit and prediction, examine model results on a graphical display.
}
\usage{
\method{plot}{GroAgeMod}(x, ...)
}
\arguments{
  \item{x}{
An object of class 'GroAgeMod' created by function GroAgeFit().
}
  \item{\dots}{
Not used.
}
}
\value{
A multipanel plot.
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples in help pages of function AgeGroFit()
}
\keyword{ ~iplot }